Update NFT Page and add .emv
